requirements.txt
boto3
paramiko
uuid
apache-airflow==1.10.0
hmsclient